Terminate analysis if destination and source operands in MIR are not known (#5)
During taint analysis, if we are unable to determine the src and dest operands in MIR, we cannot propagate the taint. To be conservative in such cases, if any of the machine operands in the MIR are present in the taint list, terminate the taint analysis with appropriate error message.